+++ server/controllers/UserController.ts
@@ -2,8 +2,8 @@ import type { Context } from 'hono';
 import { deleteCookie, getCookie } from 'hono/cookie';
 
 import db from '../database/pool';
-import { eq, and } from 'drizzle-orm';
-import { blockedList, users, followingLogs, notifications, posts } from '../database/schema';
+import { eq, and, ilike, desc } from 'drizzle-orm';
+import { blockedList, users, followingLogs, notifications, posts, reportLogs } from '../database/schema';
 
 import { IsImage, FileExist } from '../utilities/FileReader';
 import { OnProfileChange, OnUsernameChange, OnNameChange, OnEmailChange, OnPasswordChange, OnEnableSecurity } from './helpers/user';
@@ -44,7 +44,7 @@ export default class UserController {
             const user = c.get('user');
             const username: string = c.req.param('username')?.toLowerCase();
 
-            const selfDB = await db.query.users.findFirst({ where: eq(users.id, +user.id || 0) });
+            const selfDB = await db.query.users.findFirst({ where: eq(users.id, +user.id || -1) });
             if (!selfDB)
                 return c.json({ error: 'Forbidden access.' }, HTTPStatus.FORBIDDEN);
 
@@ -103,6 +103,37 @@ export default class UserController {
         }
     }
 
+    static async GetUsers(c: Context) {
+        try {
+            const user = c.get('user');
+            const username: string = c.req.param('username');
+
+            if (!user)
+                return c.json({ message: "Forbidden access." }, HTTPStatus.FORBIDDEN);
+
+            const usn = username?.trim()?.toLowerCase();
+            if (usn) {
+                const data = await db.select({
+                    id: users.id, role_id: users.role_id,
+                    full_name: users.full_name, username: users.username,
+                    profileUrl: users.profileUrl, verified: users.verified
+                }).from(users).where(ilike(users.username, `%${usn}%`)).orderBy(desc(users.followers)).limit(20);
+
+                return c.json({ users: data }, HTTPStatus.OK);
+            } else {
+                const data = await db.select({
+                    full_name: users.full_name, username: users.username,
+                    profileUrl: users.profileUrl, verified: users.verified
+                }).from(users).orderBy(desc(users.followers)).limit(20);
+
+                return c.json({ users: data }, HTTPStatus.OK);
+            }
+        } catch (err) {
+            console.error(err);
+            return c.json({ message: "Internal server error." }, HTTPStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
     static async UpdateUser(c: Context) {
         try {
             const user = c.get('user');
@@ -111,7 +142,7 @@ export default class UserController {
             const { type, id } = c.req.param();
 
             if (!user)
-                return c.json({ error: 'Forbidden access.' }, HTTPStatus.FORBIDDEN);
+                return c.json({ message: 'Forbidden access.' }, HTTPStatus.FORBIDDEN);
 
             const userId: number = +id;
             const security_code: string = body.security_code;
@@ -164,24 +195,43 @@ export default class UserController {
                 return c.json({ message: "Forbidden." }, HTTPStatus.FORBIDDEN);
 
             const result = await db.select({
-                id: notifications.id,
-                action: notifications.action,
-                createdAt: notifications.createdAt,
-                postId: notifications.postId,
-                commentId: notifications.commentId,
-                isRead: notifications.isRead,
-                postAttachment: posts.imageUrl,
-                postDescription: posts.description,
-                postOwner: posts.ownerId,
-                username: users.username,
-                verified: users.verified,
-                profileUrl: users.profileUrl
+                id: notifications.id, action: notifications.action, createdAt: notifications.createdAt,
+                postId: notifications.postId, commentId: notifications.commentId, isRead: notifications.isRead,
+                postAttachment: posts.imageUrl, postDescription: posts.description, postOwner: posts.ownerId,
+                username: users.username, verified: users.verified, profileUrl: users.profileUrl
             }).from(notifications)
                 .leftJoin(users, eq(notifications.senderId, users.id))
                 .leftJoin(posts, eq(notifications.postId, posts.id))
                 .where(eq(notifications.recipientId, +user.id));
 
             return c.json({ notifications: result }, HTTPStatus.OK);
+        } catch (err) {
+            console.log(err);
+            return c.json({ message: "Internal server error." }, HTTPStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    static async RequestVerification(c: Context) {
+        try {
+            const user = c.get("user");
+            const self = await db.query.users.findFirst({ where: eq(users.id, +user.id || -1) });
+            if (!self)
+                return c.json({ message: "Forbidden." }, HTTPStatus.FORBIDDEN);
+
+            if (self.followers! >= 1000) {
+                const where = and(eq(reportLogs.reporterId, self.id), eq(reportLogs.reporterId, self.id), eq(reportLogs.logType, "Verification requested"));
+                const existingReport = await db.query.reportLogs.findFirst({ where });
+                if (!existingReport) {
+                    await db.delete(reportLogs).where(where);
+                    await db.insert(reportLogs).values({ logType: "Verification requested", reporterId: self.id, userId: self.id, reason: "", approved: 0 });
+                }
+                else
+                    await db.insert(reportLogs).values({ logType: "Verification requested", reporterId: self.id, userId: self.id, reason: "", approved: 0 });
+
+                return c.json({ message: "Request sent." }, HTTPStatus.OK);
+            } else {
+                return c.json({ message: "You're not eligible to make a request." }, HTTPStatus.CONFLICT);
+            }
         } catch (err) {
             console.log(err);
             return c.json({ message: "Internal server error." }, HTTPStatus.INTERNAL_SERVER_ERROR);