+++ server/routes/user/User.ts
@@ -12,6 +12,7 @@ UserRouter.get('/user', Authorization, (c) => {
 
 UserRouter.get('/resync-user', Authorization, ResyncAuth);
 
+UserRouter.get('/get-users/:username?', Authorization, UserController.GetUsers);
 UserRouter.get('/user-data', Authorization, UserController.GetTokenBase);
 UserRouter.get('/user-info/', Authorization, (c) => UserController.GetProfileDB(c, true));
 UserRouter.get('/user-info/:username', Authorization, (c) => UserController.GetProfileDB(c, false));
@@ -21,6 +22,7 @@ UserRouter.get('get-notifications', Authorization, UserController.GetNotificatio
 
 UserRouter.patch('/update-data/:type/:id', Authorization, UserController.UpdateUser);
 
+UserRouter.post('/send-verification-request', Authorization, UserController.RequestVerification);
 UserRouter.post('/logout', UserController.Logout);
 
 export default UserRouter;